+++ src/main/java/ObstacleSpawn.java
@@ -2,11 +2,11 @@ package main.java;
 
 public class ObstacleSpawn {
     private String name;
-    private double speed;
+    private int speed;
     private int y;
     private double delta;
 
-    public ObstacleSpawn(String name, double speed, int y, double delta) {
+    public ObstacleSpawn(String name, int speed, int y, double delta) {
         this.name = name;
         this.speed = speed;
         this.y = y;
@@ -14,7 +14,7 @@ public class ObstacleSpawn {
     }
 
     public String getName() { return this.name; }
-    public double getSpeed() { return this.speed; }
+    public int getSpeed() { return this.speed; }
     public int getY() { return this.y; }
     public double getDelta() { return this.delta; }
 }
